requirements.txt
pandas
numpy
google-api-python-client
python-dotenv
nltk
scikit-learn
# -e .